Add sample program for LED and capactive touch.
This sample program blinks the green and red LEDs (with differing
patterns, so as to identify they're around the right way: once for
green, twice for red), and responds to capacitve touch by blinking that
LED ten times.

Compiles to flash offset 0x4000, and is thus compatible with the EFM32
and EFM32HG Silicon Labs bootloaders.