Add (untested) support for UART output.
I've not had time to test this nor will I anytime soon (I'm going on
vacation in a few days), but I'm quite confident that this will work
fine. The T and R pins on the board are wired to LEUART0 location 1, and
the Gecko SDK's retargetserial driver seems to do all the heavy lifting
capturing printf() output and writing it out the desired UART (see
retargetserialconfig.h for implementation details).