Fix CROSS_COMPILE when using ccache
If you use ccache, then the path to arm-none-eabi-gcc wil lbe
/usr/lib/ccache/arm-none-eabi-gcc

The code to find CROSS_COMPILE already assumes that arm-none-eabi-gcc
is in your PATH.  It also assumes that arm-none-eabi-objcopy and
arm-none-eabi-objdump are in the same place.  This is not true: the
latter are both in /usr/bin (or somewhere else) in your PATH.

So rather than do a complex regex, just hardcode the CROSS_COMPILE
variable.