Explicitly disable gnu build id during compilation
On some platforms the gnu build id is enabled by default, and this causes
an extra section to be added to the resulting elf. This offsets the boot
loader by 0x24 bytes, which causes the flashed image to fail.

By default pass the -Wl,--build-id=none flags to the linker to prevent
this issue.